requirements.txt
flask
whitenoise
gunicorn
opencv-python
colorthief>=0.2.1
Pillow>=8.3.1
webcolors